--- app/src/main/java/com/dmonzonis/nookpendium/AboutActivity.kt
@@ -1,15 +1,0 @@
-package com.dmonzonis.nookpendium
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.text.method.LinkMovementMethod
-import kotlinx.android.synthetic.main.activity_about.*
-
-class AboutActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_about)
-        legal_text.movementMethod = LinkMovementMethod()
-    }
-}
