+++ app/src/main/java/com/dmonzonis/nookpendium/AboutFragment.kt
@@ -1,0 +1,25 @@
+package com.dmonzonis.nookpendium
+
+import android.os.Bundle
+import android.text.method.LinkMovementMethod
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import kotlinx.android.synthetic.main.fragment_about.*
+
+class AboutFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_about, container, false)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        legal_text?.movementMethod = LinkMovementMethod()
+        super.onCreate(savedInstanceState)
+    }
+}
